refactor : 헤더 레이아웃 관련 컴포넌트 수정 및 이동 #46
