+++ src/components/layout/Header.tsx
@@ -1,0 +1,34 @@
+'use client';
+
+import React from 'react';
+import Link from 'next/link';
+import MoviesSearch from '@/components/layout/MoviesSearch';
+import CategoryDropdown from '@/components/layout/CategoryDropdown';
+import UserMenu from '@/components/layout/UserMenu';
+
+const Header = () => {
+  return (
+    <header className='sticky top-0 z-50 flex bg-slate-800 p-4'>
+      <nav className='mx-auto flex w-full max-w-[1200px] items-center justify-between gap-10 px-8'>
+        {/* 로고, 카테고리 */}
+        <section className='flex items-center gap-8 text-slate-50'>
+          {/* 로고 */}
+          <Link href={'/home'}>
+            <h1 className='text-3xl font-bold'>Logo</h1>
+          </Link>
+
+          {/* 카테고리 메뉴 컴포넌트 */}
+          <CategoryDropdown />
+        </section>
+
+        {/* 검색영역 */}
+        <MoviesSearch />
+
+        {/* 로그인 여부에 따른 메뉴 분기 컴포넌트*/}
+        <UserMenu />
+      </nav>
+    </header>
+  );
+};
+
+export default Header;
