+++ src/components/layout/HeaderLayout.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import Header from '@/components/commons/Header';
+import Header from '@/components/layout/Header';
 import { usePathname } from 'next/navigation';
 
 const HeaderLayout = () => {
